🚧 Created CoreData storage for storing of the last fetched airports locations

+++ WeatherAppGG/WeatherAppGG/Models/CurrentCondition.swift
@@ -41,10 +41,10 @@ struct CurrentCondition: Codable {
     }
     
     struct Visibility: Codable {
-        let prevailingVisDistanceQualifier: Int?
-        let distanceSm: Int?
+        let prevailingVisDistanceQualifier: Double?
+        let distanceSm: Double?
         let distanceQualifier: Int?
-        let prevailingVisSm: Int?
+        let prevailingVisSm: Double?
     }
     
     struct Wind: Codable {
